TG-1097 Refactor all components to use css modules and remove styled-component dependency
