+++ {{cookiecutter.project_dirname}}/__tests__/components/Navbar.test.tsx
@@ -1,13 +1,11 @@
 import { expect } from '@jest/globals'
 import { screen } from '@testing-library/react'
+import { render } from '@testing-library/react'
 import React from 'react'
 
 import { Navbar } from '@/components/Navbar'
-import { renderWithWrappers } from '../functions'
-
-const setup = () => renderWithWrappers(<Navbar />)
 
 test('Navbar renders correctly', () => {
-  setup()
+  render(<Navbar />)
   expect(screen.getByRole('img')).toBeTruthy()
 })
