+++ {{cookiecutter.project_dirname}}/next.config.js
@@ -2,27 +2,22 @@
 
 const { withSentryConfig } = require('@sentry/nextjs')
 
-const {
-  SENTRY_AUTH_TOKEN,
-  SENTRY_ORG,
-  SENTRY_PROJECT_NAME,
-  SENTRY_URL,
-} = process.env
+const { SENTRY_AUTH_TOKEN, SENTRY_ORG, SENTRY_PROJECT_NAME, SENTRY_URL } =
+  process.env
 
 const nextjsConfig = {
-  compiler: { styledComponents: true },
-  i18n:{
+  i18n: {
     defaultLocale: 'en-US',
-    locales: ['en-US', 'it-IT'],
+    locales: ['en-US', 'it-IT']
   },
   output: 'standalone',
   reactStrictMode: true,
   swcMinify: true,
   rewrites: async () => [
     {
-      source: '/{{ cookiecutter.service_slug }}/health',
+      source: '/frontend/health',
       destination: '/api/health'
-    },
+    }
   ]
 }
 
@@ -37,8 +32,9 @@ const SentryWebpackPluginOptions = {
 
 // Make sure adding Sentry options is the last code to run before exporting, to
 // ensure that your source maps include changes from all other Webpack plugins
-const config = SENTRY_AUTH_TOKEN && SENTRY_ORG && SENTRY_URL
-  ? withSentryConfig(nextjsConfig, SentryWebpackPluginOptions)
-  : nextjsConfig
+const config =
+  SENTRY_AUTH_TOKEN && SENTRY_ORG && SENTRY_URL
+    ? withSentryConfig(nextjsConfig, SentryWebpackPluginOptions)
+    : nextjsConfig
 
 module.exports = config
