+++ {{cookiecutter.project_dirname}}/src/components/Navbar.tsx
@@ -1,29 +1,19 @@
 import React from 'react'
-import styled from 'styled-components'
+import styles from '@/styles/navbar.module.css'
 
 import Image from 'next/image'
 
 const Navbar = () => {
   return (
-    <Nav>
+    <nav className={styles.navbar}>
       <Image
         src={`/images/logo-light.svg`}
         width='110'
         height='48'
         alt={'sito logo'}
       />
-    </Nav>
+    </nav>
   )
 }
-
-const Nav = styled.nav`
-  width: 100%;
-  background-color: ${({ theme }) => theme.colors.neutrals[100]};
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  padding: 5px;
-`
 
 export { Navbar }
