+++ {{cookiecutter.project_dirname}}/src/components/layout/Base.tsx
@@ -1,9 +1,24 @@
-import { ThemeProvider } from 'styled-components'
 import React from 'react'
+import {
+  Open_Sans as OpenSans,
+  Hepta_Slab as HeptaSlab
+} from '@next/font/google'
 
-import { GlobalStyle } from '@/styles/GlobalStyle'
 import { Navbar } from '@/components/Navbar'
-import themes from '@/styles/themes'
+
+const baseFont = OpenSans({
+  subsets: ['latin'],
+  weight: ['400', '500', '700'],
+  style: ['normal', 'italic'],
+  variable: '--font-base'
+})
+
+const titleFont = HeptaSlab({
+  subsets: ['latin'],
+  weight: ['400', '500', '700'],
+  style: ['normal'],
+  variable: '--font-title'
+})
 
 type Props = {
   children: React.ReactNode
@@ -11,11 +26,10 @@ type Props = {
 
 const Layout = ({ children }: Props) => {
   return (
-    <ThemeProvider theme={themes['light']}>
-      <GlobalStyle />
+    <div className={`${titleFont.variable} ${baseFont.variable}`}>
       <Navbar />
       {children}
-    </ThemeProvider>
+    </div>
   )
 }
