+++ {{cookiecutter.project_dirname}}/src/pages/_document.tsx
@@ -1,36 +1,7 @@
 import React from 'react'
-import { ServerStyleSheet } from 'styled-components'
 import Document, { Html, Head, Main, NextScript } from 'next/document'
 
-import type { DocumentInitialProps, DocumentContext } from 'next/document'
-
 export default class MyDocument extends Document {
-  static async getInitialProps(
-    ctx: DocumentContext
-  ): Promise<DocumentInitialProps> {
-    const sheet = new ServerStyleSheet()
-    const originalRenderPage = ctx.renderPage
-
-    try {
-      ctx.renderPage = () =>
-        originalRenderPage({
-          enhanceApp: App =>
-            function renderPage(props) {
-              return sheet.collectStyles(<App {...props} />)
-            }
-        })
-
-      const initialProps = await Document.getInitialProps(ctx)
-
-      return {
-        ...initialProps,
-        styles: [initialProps.styles, sheet.getStyleElement()]
-      }
-    } finally {
-      sheet.seal()
-    }
-  }
-
   render() {
     const { locale } = this.props
     return (
