+++ {{cookiecutter.project_dirname}}/src/pages/index.tsx
@@ -1,31 +1,14 @@
 import React from 'react'
-import styled from 'styled-components'
+import styles from '@/styles/index.module.css'
 
 import type { NextPage } from 'next'
 
 const Home: NextPage = () => {
   return (
-    <Container>
-      <H1>Hello World!</H1>
-    </Container>
+    <div className={styles.container}>
+      <h1 className={styles.title}>Hello World!</h1>
+    </div>
   )
 }
-
-const H1 = styled.h1`
-  font-size: 40px;
-  font-weight: 500;
-`
-
-const Container = styled.div`
-  display: flex;
-  flex: 1;
-  min-height: calc(100vh - 80px);
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  button {
-    margin-top: 30px;
-  }
-`
 
 export default Home
