--- {{cookiecutter.project_dirname}}/src/styles/themes/index.ts
@@ -1,14 +1,0 @@
-import { theme as light } from '@/styles/themes/light'
-import { theme as dark } from '@/styles/themes/dark'
-
-export enum Theme {
-  light = 'light',
-  dark = 'dark'
-}
-
-const theme = {
-  [Theme.light]: light,
-  [Theme.dark]: dark
-}
-
-export default theme
